Added the field limits to the urls.
* Why was this change necessary?
The pebble app is mutch faster if we do only get this info we really use.
* How does it address the problem?
The public transport api provides a mechanism to limit the fields. This can be
done by using the url parameter fields[].
* Are there any side effects?
I tested it with my hw app and the emulator. It was really faster in loading
the data. I did not see any negativ side effects.

https://github.com/D-32/PebbleSBB/issues/2